Rewrite timer notification sfx to exec paplay.
I ran into the "Warning: Jack Server not found" when testing the QML
Audio on my TV laptop that has the laptop screen disabled and the HDMI
screen as primary.

Also, since Chrome (or was it Firefox) would always default the audio
to the laptop speakers after a certain browser update, I disabled the
laptop speaker Sink/Output so it would always default to the HDMI
Sink/Output.

It seems the Qt/QML audio code probably doesn't take this edge case
into account and probably tries to fallback to looking for a JACK
audio server. When it can't find one, it just spits out that it can't
find one.

https://github.com/Zren/plasma-applet-eventcalendar/issues/27